Refactored Link component to make it reusable in multiple components
